refactor: changed the structure of sign in and sign up components/pages
Now there are simply two pages, SignIn and SignUp, which share the same CSS through the imported formStyles.ts. This was done to make the code more reusable
