+++ client/src/pages/Cart.tsx
@@ -11,10 +11,10 @@ import { Fragment, useContext } from 'react';
 import { Link } from 'react-router-dom';
 import CartProduct from '../components/CartProduct';
 import CheckoutForm from '../components/CheckoutForm';
-import { SignInForm } from '../components/SignInForm';
 import { useAuth } from '../contexts/AuthContext';
 import { ProductContext } from '../contexts/ProductContext';
 import { useShoppingCart } from '../contexts/ShoppingCartContext';
+import { SignInForm } from './SignIn';
 
 function Cart() {
   const { cartProducts, cartQuantity } = useShoppingCart();
